Refactor focus timer class
